requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0